Support fractional numbers of pizza

+++ pizzaSharing.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-pizzas = int(input("How many pizzas do you have? "))
+pizzas = float(input("How many pizzas do you have? "))
 students = int(input("How many students are there? "))
 
 print(f"Each student should eat {pizzas/students} pizzas")
